// src/app/infrastructure/dto/response/dataTaxResponse.dto.ts
export interface DataTaxResponseDto {
  Item: number;
  IDPremio: number;
  IDCatalogo: number;
  ValorUnitario: number;
  Cantidad: number;
  PorcentajeImpuesto: number;
  SubTotalItem: number;
  ValorBase: number;
  ValorImpuesto: number;
  ValorComision: number;
  ImpuestoComision: number;
  PuntosEquivalentes: number;
  PuntosUnidad: number;
  Total: number;
}